--- .github/workflows/build.yml
@@ -1,22 +1,0 @@
-name: Build
-
-on: [push]
-
-jobs:
-  build:
-
-    runs-on: windows-latest
-
-    steps:
-    - uses: actions/checkout@v2
-    - name: set up JDK 1.8
-      uses: actions/setup-java@v1
-      with:
-        java-version: 1.8
-    - name: Build with Gradle
-      run: ./gradlew build
-    - name: Upload artifact
-      uses: actions/upload-artifact@v1.0.0
-      with:
-        name: PassableFoliage
-        path: build/libs
