+++ build.gradle
@@ -188,9 +188,9 @@ dependencies {
     minecraft "net.minecraftforge:forge:${minecraft_version}-${forge_version}"
     annotationProcessor "org.spongepowered:mixin:${mixin_version}:processor"
     
-    implementation fg.deobf("curse.maven:jade-324717:4318720")
-    implementation fg.deobf("curse.maven:kiwi-303657:4287363")
-    annotationProcessor "curse.maven:kiwi-303657:4287363"
+    implementation fg.deobf("curse.maven:jade-324717:4439089")
+    implementation fg.deobf("curse.maven:kiwi-303657:4500472")
+    annotationProcessor "curse.maven:kiwi-303657:4500472"
 	
     implementation fg.deobf("me.shedaniel.cloth:cloth-config-forge:${project.cloth_config_version}")
 }
@@ -239,7 +239,7 @@ modrinth {
     versionNumber = project.mod_version // You don't need to set this manually. Will fail if Modrinth has this version already
     versionType = "release" // This is the default -- can also be `beta` or `alpha`
     uploadFile = jar // With Loom, this MUST be set to `remapJar` instead of `jar`!
-    gameVersions = ["1.19.3"] // Must be an array, even with only one version
+    gameVersions = ["1.19.4"] // Must be an array, even with only one version
     loaders = ["forge"] // Must also be an array - no need to specify this if you're using Loom or ForgeGradle
     dependencies { // A special DSL for creating dependencies
         // scope.type
