+++ gradle.properties
@@ -3,17 +3,17 @@
 org.gradle.jvmargs=-Xmx3G
 org.gradle.daemon=false
 
-minecraft_version=1.19.3
-forge_version=44.1.5
+minecraft_version=1.19.4
+forge_version=45.0.49
 mappings_channel=official
-mappings_version=1.19.3
+mappings_version=1.19.4
 
 mod_id=passablefoliage
 mod_name=PassableFoliage
-mod_version=6.1.0
+mod_version=7.0.0
 mod_base_package=snownee.passablefoliage
 
 # Dependencies
 mixin_version=0.8.5
 jei_version=8.0.0.14
-cloth_config_version=9.0.94
+cloth_config_version=10.0.96
