+++ src/main/resources/assets/passablefoliage/lang/en_us.json
@@ -1,4 +1,14 @@
 {
-    "enchantment.passablefoliage.leaf_walker": "Leaf Walker",
-    "enchantment.passablefoliage.leaf_walker.desc": "Allows the player to walk on foliage blocks."
+  "enchantment.passablefoliage.leaf_walker": "Leaf Walker",
+  "enchantment.passablefoliage.leaf_walker.desc": "Allows the player to walk on foliage blocks.",
+  "passablefoliage.config.fallDamageReduction.desc": "The percentage of normal damage taken when taking damage from falling into leaves",
+  "passablefoliage.config.fallDamageThreshold.desc": "When falling into leaves, the (block) distance a player or mob has to fall before taking damage",
+  "passablefoliage.config.speedReductionHorizontal.desc": "The reduced horizontal speed when passing through leaves. (% of normal)",
+  "passablefoliage.config.speedReductionVertical.desc": "The reduced vertical speed when passing through leaves. (% of normal)",
+  "passablefoliage.config.modifyPathFinding.desc": "Should entities recognize leaves as air",
+  "passablefoliage.config.playerOnly.desc": "Should only players be able to pass through leaves",
+  "passablefoliage.config.alwaysNotViewBlocking.desc": "Prevents leaves from rendering a texture in front of the player's view",
+  "passablefoliage.config.alwaysLeafWalking.desc": "Always consider the player to have the Leaf Walker enchantment.\nOnce enabled, the recipe of Leaf Walker enchantment will be removed from the game, and the fall damage reduction and speed reduction will be ignored.",
+  "passablefoliage.config.sounds.playerOnly.desc": "Only players can make noises",
+  "passablefoliage.config.sounds.volume.desc": "The volume of the sound made when moving inside leaves"
 }
